Fix warnings in tests, gcc-8

+++ tests/base/function_parser_04.cc
@@ -83,7 +83,7 @@ test2()
                   copy_data());
   std::cout << "result: " << result << std::endl;
 
-  Assert(result == v.size(), ExcMessage("uhuh!"));
+  Assert((unsigned int)result == v.size(), ExcMessage("uhuh!"));
 }
